feat: settings page multi-column grid layout (#131)
Replace the single-column settings page with a responsive grid of
navigable link cards grouped by category (Integrations, Network,
System). Each settings section is now its own sub-page, keeping
forms focused and the overview scannable.

- 1 col mobile, 2 col md, 3 col lg via Tailwind grid
- Moved VyOS, NPM, Scanner, Webhook, Retention, Speed Test, and
  Password forms into dedicated /settings/* sub-pages
- Updated TopBar password link to route to /settings/password

Closes #131
